backend/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
pyyaml==5.4.1
uritemplate==3.0.1

djangorestframework-simplejwt==4.8.0
drf-yasg==1.20.0

gunicorn==20.1.0
psycopg2==2.9.1

flake8==3.9.2
pre-commit==2.14.0
pytest-django==4.4.0
